Team page, navbar improvements
Improved the "Meet the team" page. Added a licenses.html which I will fill out with software licenses later.
I removed an incorrect portion of the web/readme.md, and added the GitHub logo to the navbar. I also changed the navbar to a drop-down menu to save space. Finally, I updated both templates with the jQuery and Bootstrap.min.js CDNs.