Add try/except to ingest
Some events have objectID's with lengths of over 280 characters. These are almost always malformed/improper links. I added try/except blocks so that if the script encounters one of these DOI's, it will not add it to the databse, but rather skip it and return to main.py

Main.py also prints each event, which aids in debugging.