Creation of SciELO Data Gather
This script takes PID's from SCI and converts it into DOI's that can be used in the Crossref API. We take that information and insert it into the doidata database. When we run our metadata update script, the script will see the new DOI's and then get the metadata instead of having to make a seperate metadata script.